clean up api endpoints
