+++ src/main/java/pl/cyfronet/indigo/controller/IndexController.java
@@ -3,9 +3,12 @@ package pl.cyfronet.indigo.controller;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.view.RedirectView;
 import pl.cyfronet.indigo.security.PortalUser;
 
+import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 
 @Controller
@@ -29,5 +32,14 @@ public class IndexController {
         }
         mav.getModelMap().addAttribute("version", version);
         return mav;
+    }
+
+    @RequestMapping(value = "/auth/logout", method = RequestMethod.GET)
+    public RedirectView logout(HttpSession session) {
+        session.invalidate();
+        RedirectView redirectView = new RedirectView();
+        redirectView.setContextRelative(true);
+        redirectView.setUrl("/");
+        return redirectView;
     }
 }
